+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 + `changelog`: Added changelog checker github workflow
 + `changelog`: Added support for data describe in API
 + `changelog`: Added support for timestamps with timezone offsets
++ `changelog`: Include serving endpoint information in FeatureList.info()
 
 ### 🧰 Bug fixes 🧰